TritonClient/requirements.txt
msgpack==1.0.4
redis>=4.3.4
numpy==1.23.0
opencv-python-headless==4.6.0.66
paho-mqtt
pympler
simplejpeg